+++ src/priorityQueue.h
@@ -23,7 +23,7 @@ class PriorityQueue : public GenericQueue {
   public:
     virtual Order* next() override;
     virtual void push(Order* order) override;
-    virtual double getQueueLength(bool vip) override;
+    virtual int getQueueLength(bool vip) override;
 };
 
 #endif /* PRIORITYQUEUE_H_ */
